requirements.txt
numpy==1.24.3
pillow
tensorflow==2.15.0
flask
gradio